requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.1.7
python-multipart==0.0.6
python-telegram-bot==20.6
pydantic==2.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.25.0 
sqlalchemy==2.0.15
bcrypt==4.0.1
pytest==7.3.1 
pika==1.3.2
